@tmontaigu 2023-10-12
@slab-ci cpu_integer_test