@tmontaigu 2023-10-12
@slab-ci cpu_fast_test
@slab-ci cpu_test