@IceTDrinker 2023-10-12
if we are getting sig kills we may want to change the number of cargo nextest processes we launch in parallel