@IceTDrinker 2023-10-12
or limit the tests given the biggest server key size so that it fits in RAM, as the tests are multi threaded anyways we should not be losing too much hopefully

btw the test executor taking the key by value will make things worse right ? Could we maybe move the key in the executor and have an accessor on the executor to get a ref to the server key so that we don't have an extra copy in memory ?