@tmontaigu 2023-10-12
> btw the test executor taking the key by value will make things worse right ? Could we maybe move the key in the executor and have an accessor on the executor to get a ref to the server key so that we don't have an extra copy in memory ?

Yes that should work